chore: Split setupRunnablesBeforeClosingOfApplication
